requirements.txt
colorama==0.4.5
cryptography==38.0.3
defusedxml==0.7.1
Faker==15.3.1
pyOpenSSL==23.2.0
requests==2.28.1
tldextract==3.4.0
